requirements.txt
acres==0.5.0
azure-ai-documentintelligence==1.0.2
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.34.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
dotenv==0.9.9
etelemetry==0.3.1
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
msrest==0.7.1
neo4j==5.28.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.3.0
prov==2.1.1
puremagic==1.29
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
safetensors==0.5.3
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
